soal_3/requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
fpdf==1.7.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==20.3
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
